Refactor VideoDevice into its own module
