+++ coco-vm/src/lib.rs
@@ -1,10 +1,9 @@
+mod video;
+
 use coco_core::{Cpu, Machine};
+use video::{VideoBuffer, VideoDevice};
 
-pub const SCREEN_WIDTH: usize = 192;
-pub const SCREEN_HEIGHT: usize = 144;
-const VIDEO_BUFFER_LEN: usize = SCREEN_WIDTH * SCREEN_HEIGHT;
-
-pub type Pixel = u8;
+pub use video::{SCREEN_HEIGHT, SCREEN_WIDTH};
 
 #[derive(Debug, Clone, Copy)]
 pub struct Output {
@@ -42,25 +41,7 @@ impl Vm {
         Output { pc: cpu.pc() }
     }
 
-    pub fn pixels(&self) -> (&[Pixel; VIDEO_BUFFER_LEN], &[Pixel; VIDEO_BUFFER_LEN]) {
+    pub fn pixels(&self) -> (&VideoBuffer, &VideoBuffer) {
         (&self.video.background, &self.video.foreground)
-    }
-}
-#[derive(Debug)]
-struct VideoDevice {
-    background: [Pixel; VIDEO_BUFFER_LEN],
-    foreground: [Pixel; VIDEO_BUFFER_LEN],
-}
-
-impl VideoDevice {
-    pub fn new() -> Self {
-        let mut buffer = [0x00 as Pixel; VIDEO_BUFFER_LEN];
-        for i in 0..VIDEO_BUFFER_LEN {
-            buffer[i] = (i % 0x10) as Pixel;
-        }
-        Self {
-            background: [0x00; VIDEO_BUFFER_LEN],
-            foreground: buffer,
-        }
     }
 }
